Se mueven ultimas busquedas a un componente independiente.
