+++ src/components/Profile/Profile.styles.jsx
@@ -32,5 +32,3 @@ export const InfoContainer = styled(Box)({
 });
 
 export const LastBuys = styled(Box)({});
-
-export const LastSearches = styled(Box)({});
